Moving all activity classes into a separate  package.
