+++ app/src/main/java/com/example/myapplication/Activity/WelcomeActivity.java
@@ -1,4 +1,4 @@
-package com.example.myapplication;
+package com.example.myapplication.Activity;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -9,7 +9,7 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
 
-import com.example.myapplication.Models.UserType;
+import com.example.myapplication.R;
 
 public class WelcomeActivity extends AppCompatActivity {
