+++ app/src/main/res/layout/activity_answer_question.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="com.example.myapplication.AnswerQuestionActivity">
+    tools:context="com.example.myapplication.Activity.AnswerQuestionActivity">
 
 
     <TextView
