+++ app/src/main/res/layout/activity_create_question.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="com.example.myapplication.CreateQuestionActivity">
+    tools:context="com.example.myapplication.Activity.CreateQuestionActivity">
 
     <Button
         android:layout_width="wrap_content"
