+++ app/src/main/res/layout/activity_create_quiz.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CreateQuizActivity">
+    tools:context=".Activity.CreateQuizActivity">
 
     <TextView
         android:id="@+id/createQuizActivity_headerMessage"
