+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="com.example.myapplication.MainActivity">
+    tools:context="com.example.myapplication.Activity.MainActivity">
 
     <Button
         android:layout_width="wrap_content"
