+++ app/src/main/res/layout/activity_welcome.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".WelcomeActivity">
+    tools:context=".Activity.WelcomeActivity">
 
     <TextView
         android:id="@+id/welcomeActivity_bannerMessage"
